haystack-samples/pyproject.toml
[tool.poetry]
name = "haystack-samples"
version = "0.1.0"
description = ""
authors = ["Gavin Noronha <gavinln@hotmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
haystack-ai = "^2.2.1"
trafilatura = "^1.10.0"
ipython = "^8.25.0"
fire = "^0.6.0"
chromadb = "^0.5.0"
chroma-haystack = "^0.18.0"
datasets = "^2.20.0"
torch = {version = "^2.3.1+cpu", source = "torch-cpu"}
sentence-transformers = "^3.0.1"
markdown-it-py = "^3.0.0"
mdit-plain = "^1.0.1"
pypdf = "^4.2.0"
wikipedia = "^1.4.0"
pyyaml = "^6.0.1"
transformers = {extras = ["sentencepiece", "torch"], version = "^4.41.2"}
colorama = "^0.4.6"

[[tool.poetry.source]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[[tool.poetry.source]]
name = "torch-gpu"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
